fix(client/context/auth): Fix eslint error
