+++ client/src/contexts/AuthContext.ts
@@ -1,0 +1,12 @@
+import { createContext } from 'react';
+import { User } from '@/types/graphql-generated';
+
+export type AuthContextType = {
+  user: User | null;
+  isAuthenticated: boolean;
+  isLoading: boolean;
+  error: string | null;
+  login: (email: string, password: string) => Promise<void>;
+};
+
+export const AuthContext = createContext<AuthContextType | undefined>(undefined);
